some improvments
these improvements make the multiplay script.  It does not do anything new at the core.  However, it does let the use know what is going on in more detail and offers the options for them to exit the script without using ctrl+c if they realize that they don't actually want to proceed.